Set ACL on S3 uploads to 'public-read'
Without this, uploads will assume the bucket's default acl, which,
depending on your configuration, may not be public. If the bucket is
not set up as a static website, after saving, the uploader will hang
since the uploaded file will not be available to view.

However, since this storage module is specifically for making
resources public, this adds the public read acl for uploaded resources
so that they'll be visible regardless of the permissions on the
bucket they're uploaded to.